backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
pydantic==2.5.0
python-multipart==0.0.6
redis==5.0.1
rq==1.15.1
requests==2.31.0
python-dotenv==1.0.0
pillow==10.1.0
opencv-python==4.8.1.78
ultralytics==8.0.206
torch==2.1.1
torchvision==0.16.1
easyocr==1.7.0
imagehash==4.3.1
instaloader==4.10.3
selenium==4.15.2
webdriver-manager==4.0.1
geopy==2.4.0
structlog==23.2.0
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
websockets==12.0